+++ crates/fuel-core/src/query/subscriptions/test.rs
@@ -55,6 +55,7 @@ fn success() -> TransactionStatus {
         block_id: Default::default(),
         time: Tai64(0),
         result: None,
+        receipts: vec![],
     }
 }
 
@@ -64,7 +65,7 @@ fn failed() -> TransactionStatus {
         block_id: Default::default(),
         time: Tai64(0),
         result: None,
-        reason: Default::default(),
+        receipts: vec![],
     }
 }